@@ -2,7 +2,7 @@ name: android_alarm_manager_example
 description: Demonstrates how to use the android_alarm_manager_plus plugin.
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
 
 dependencies:
@@ -10,15 +10,16 @@ dependencies:
     sdk: flutter
   android_alarm_manager_plus:
     path: ../
-  shared_preferences: ^0.5.6
-  e2e: 0.3.0
-  path_provider: ^1.3.1
+  shared_preferences: ^2.0.12
+  path_provider: ^2.0.8
 
 dev_dependencies:
   espresso: ^0.0.1+5
   flutter_driver:
     sdk: flutter
   flutter_test:
+    sdk: flutter
+  integration_test:
     sdk: flutter
   flutter_lints: ^1.0.4
 
